+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0 (January 26, 2017)
+* Refactored multiyargs to return a promise.
+* Replaced `jshint` by `standard`.
+
 ## 0.1.2 (January 19, 2017)
 * Enhancement: add optional `forEachCallback` as third argument.